caffeine-meta
This is the meta repository for all things relating to our projects. There isn't much in this repository, so you should take a look at this issue tracker instead.

Want to help?
Issues labeled with "good first issue" or "help wanted" are great places to start. Right now, we're still in the process of setting things up, so if you'd like to register your interest in helping and can commit some of your time to it, consider leaving a comment.

When will a task be completed?
I'm asking myself this question every day. Some things get blocked on me (JellySquid) and this can result in slowness. Other times, it's a matter of finding the resources for something, as things like servers require more investment. Unless you are involved with the task, please try to avoid asking for updates -- it stretches my already thin resources out further.

Can anyone open an issue?
If you think you have a meta issue about the projects, our website, build pipelines, or anything else, feel free to. If your issue is specific to one project, it likely isn't suitable for this issue tracker. You should not be reporting bugs, crashes, or anything like that to this repository.